Projeto funcionando tudo como prometido nesse primeiro Sprint. Cadastrar novo Player: Podem ser cadastrados vários perfis diferentes, e vc decide com qual Player vai jogar naquele momento; Segurança: De acordo com cada Player o jogo é capaz de recuperar as informações, como Itens do Player, Habilidades e Personagens comprados/desbloqueados; Inventário/Loja: Temos a tela que mostra qual o hérois que eu acabei de adquirir, e a tela que mostra quais heróis eu ainda posso adquirir. Também é possível ver nessas telas os valores em Gemas que o Player tem.

+++ src/br/com/seven/lsi/model/Personagem.java
@@ -28,7 +28,7 @@ public class Personagem implements Serializable {
     private Double vida;
     private Double defesa;
     private Double ataque;
-    private Double valorVenda;
+    private int valorVenda;
     @OneToMany
     private List<Habilidade> habilidades;
     @OneToMany
@@ -98,11 +98,11 @@ public class Personagem implements Serializable {
         this.itens = itens;
     }
 
-    public Double getValorVenda() {
+    public int getValorVenda() {
         return valorVenda;
     }
 
-    public void setValorVenda(Double valorVenda) {
+    public void setValorVenda(int valorVenda) {
         this.valorVenda = valorVenda;
-    }    
+    }
 }
